fix(Demuxer2): 修改队列长度, 修复某些情况下seek会卡主

+++ include/helper.h
@@ -33,7 +33,7 @@ extern "C" {
 }
 #pragma GCC diagnostic ignored "-Wold-style-cast"
 const int ERROR_EOF = AVERROR_EOF;
-const size_t MAX_AUDIO_FRAME_SIZE = 192000;
+const int MAX_AUDIO_FRAME_SIZE = 192000;
 #pragma GCC diagnostic pop
 
 #endif //PONYPLAYER_HELPER_H
